backend/requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
psycopg2-binary
SQLAlchemy
Flask-Migrate
flask-smorest
marshmallow
apispec
flask-cors
google-generativeai
requests
python-dotenv